[#10] refactor: src/apis/dto를 src/types로 이동
