--- src/apis/dto/auth.ts
@@ -1,8 +1,0 @@
-export interface LoginRequestBody {
-  userId: string;
-  password: string;
-}
-
-export interface AuthResponseBody {
-  access: string;
-}
