--- src/apis/dto/users.ts
@@ -1,8 +1,0 @@
-export interface UpdateInfoResponseBody {
-  createdAt: string;
-  updatedAt: string;
-}
-
-export interface User {
-  userId: string;
-}
