+++ src/types/auth.ts
@@ -1,0 +1,8 @@
+export type LoginRequestBody = {
+  userId: string;
+  password: string;
+};
+
+export type AuthResponseBody = {
+  access: string;
+};
