+++ src/types/users.ts
@@ -1,0 +1,8 @@
+export type UpdateInfoResponseBody = {
+  createdAt: string;
+  updatedAt: string;
+};
+
+export type User = {
+  userId: string;
+};
